libview: Make page layout a mode
Instead of having a few different mutually exclusive booleans it
makes sense to have it be a mode with the following options:
automatic, single, dual.

This allows us to have a way to automatically determine if dual page
mode should be used when the window is large enough.

https://bugzilla.gnome.org/show_bug.cgi?id=689468
origin commit:
https://git.gnome.org/browse/evince/commit/?h=gnome-3-10&id=42f6d4b