libview: Refactor code for drawing page and selection surfaces
The code necessary to draw a page surface or a selection
surface is the same. We factor this out to  a new static method
called draw_surface

origin commit:
https://git.gnome.org/browse/evince/commit/?h=gnome-3-8&id=3ab6ac1

libview: Fix rendering of the first visible page while resizing

While we are resizing the view widget and waiting for a new
surface rendered at the right size, we use old surfaces
scaled to match the target size. When the begining of a page is not
visible, the scaled surfaces are not placed correctly.

origin commit:
https://git.gnome.org/browse/evince/commit/?h=gnome-3-8&id=72f2ae4