libview: Actually update the page when moving the care from a different page
I removed the line to set the current page after moving the caret cursor
by mistake. Add it back and change cursor_go_to_next/previous_page to
not update the page, but only update the cursor_page so that we have a
single place where the page is updated after caret cursor moves.

origin commit:
https://git.gnome.org/browse/evince/commit/?h=gnome-3-10&id=bf85030