libview: Don't update current_page to negative number
In continuous mode with dual page view, there is a chance
that the pages were outside of the scroll view.  When
opening a document in such circumstances, the current
page should not be updated to a negative number when
already has the first page as default.

This fixes a crash when searching with a negative
current page.

Fix https://bugzilla.gnome.org/show_bug.cgi?id=683172
origin commit:
https://git.gnome.org/browse/evince/commit/?h=gnome-3-10&id=8fe6917