libview: merge get_caret_cursor_rect_from_offset and get_caret_cursor_area
Into a single function and use get_caret_cursor_area everywhere.
get_caret_cursor_rect_from_offset() was a bit confusing, because it
didn't return the cursor_rect, but the bbox of the character where the
cursor should be at. The method that actually returns the cursor_rect is
get_caret_cursor_area.

origin commit:
https://git.gnome.org/browse/evince/commit/?h=gnome-3-10&id=13a5978