ev-view-accessible: call document_changed_cb when setting the model
When an application embedding an EvView reuses that view to display a new
document, the content of last page viewed from the previous document is
exposed to assistive technologies. In order to get the content of the
current document, the user must change pages. This problem resulted from
the assumption that the document would be set after the model was set.

https://bugzilla.gnome.org/show_bug.cgi?id=709375

https://git.gnome.org/browse/evince/commit/?h=gnome-3-10&id=3d06f42