shell: Store and restore current folder when opening or saving file
Remember the folder last used to open or save a document, attachment
or image.

taken from:
https://git.gnome.org/browse/evince/commit/shell/ev-window.c?h=gnome-3-6&id=f6137df22503d88a76cc69f622919ec7bd3582d7

shell: Fix potential read-after-free in gsettings use
It's not valid to use '&' with g_settings_get().

https://bugzilla.gnome.org/show_bug.cgi?id=723737

taken from:
https://git.gnome.org/browse/evince/commit/?id=41ce0e1

ev-window: Don't crash when trying to save a copy

We can not free both folder_uri and parent_uri. If parent_uri is
non-NULL then it will point to the same location as folder_uri.
Also, parent_uri is now unused so we can just remove it.

https://bugzilla.gnome.org/show_bug.cgi?id=727536

taken from:
https://git.gnome.org/browse/evince/commit/?id=3850ff5