Setting parent for annotation properties dialog
After the last gtk+ update, the warning:

"Gtk-Message: GtkDialog mapped without a transient
parent. This is discouraged."

was shown whenever the annotation properties
dialog was opened. The commit fixes this issue by
setting this dialog transient for the main window.
As a plus, the dialog is now shown always on the
center of the main window.

https://bugzilla.gnome.org/show_bug.cgi?id=732114

taken from:
https://git.gnome.org/browse/evince/commit/?id=9533c68