ev-view: Add pan gesture to switch page
This only applies on non-continuous mode, and the page fits with no
horizontal scrolling. The gesture is actually connected to the parent
GtkScrolledWindow (tracked through hierarchy events), so it is able
to interact with kinetic scrolling gestures there, and cancel those
if the pan gesture is recognized.

Upstream commits by garnacho on 6 May and 15 Aug 2014:
https://github.com/GNOME/evince/commit/afc50e7e07c135e22137a33cbf9913713692bf49
https://github.com/GNOME/evince/commit/2280b09ec71ee88eeeb7fd8d02adb2985d17a765