sidebar-thumbnails: fix clunky scrolling
Caused by GtkIconView doing an invalidate and relayout of *all*
items in the view anytime we update model data in any indiviual
item (which happens with all the items that are getting in and out
of the scrolling area while we scroll). This caused GtkIconView to
machine-gunned us with "size-allocate" signals, a signal we were
using to update thumbnails when the sidebar is resized.

Fixed by connecting to the GtkTreeModel "row-changed" signal before
GtkIconView does it, and stop emission from there.

As we don't depend now on "size-allocate" signals to show thumbnails
while we scroll, just queue a draw on the icon view when a
thumbnail finish rendering.

Thanks Jose Aliste for first spotting the problem.

https://bugzilla.gnome.org/show_bug.cgi?id=691448
origin commit:
https://git.gnome.org/browse/evince/commit/?id=6480c70

Fixes https://github.com/mate-desktop/atril/issues/255