annotations-sidebar: Clean up the styling
Not sure why we need a GtkNotebook with a tab dedicated to nothing
but holding a single button. Ditch the notebook and put the add button
at the bottom of the sidebar.

The 'add annontation button' is edited for Atril,
just a GtkStock repacement without using GtkStock ;-)
And we use a GtkButtonBox instead of a GtkToolbar as parent for the
'add button', to be consistent with the other sidebars.

inspired by:
https://github.com/linuxmint/xreader/commit/b8627de