Save/restore context when getting colors for a different state
Getting colors for a state different from the current state of the
corresponding widget without saving the context might trigger an
invalidation and a redraw. Because this was happening from the draw
function this resulted in a redraw loop and constant high CPU usage.

This could be triggered by selecting text or searching.

https://bugzilla.gnome.org/show_bug.cgi?id=755442

inspired by:
https://git.gnome.org/browse/evince/commit/?id=1987f04
https://github.com/linuxmint/xreader/commit/e76a18d