libview: use css to draw the background of presentations
With the recent changes in gtk+, widgets have to draw themselves,
causing the current use of gdk_window_set_background_rgba to fail.

https://bugzilla.gnome.org/show_bug.cgi?id=765557

view: Redraw ev-view-presentation when setting normal and black mode.

The black and normal mode are the same from the CSS point of view.
The difference is that in the draw function the page is not drawn
in black mode. Hence, we need to explicitly queue a redraw in these
cases. Since setting the white mode add a CSS class, this queues
the redraw for us.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=772390.

origin commits:
https://git.gnome.org/browse/evince/commit/?id=2b352b3
https://git.gnome.org/browse/evince/commit/?id=b3f49f4

Fixes https://github.com/mate-desktop/atril/issues/232