Fix tests with AT-SPI2 >= 2.53
Role name for push buttons changed from "push button" to "button" for
uniformizing APIs, but that means the names Dogtail use have to match,
as it's not using the role IDs but names.

Fix this by trying to ask AT-SPI2 for the name, and falling back on the
historical name in the odd case it'd fail.