epub: Try and locate MathJax at configure time
Based on an XReader change[1], look up the MathJax source at configure
time for better portability.  If not specified manually, look in the
common locations.

This adds the --with-mathjax-directory=DIR configure option to manually
specify the location, overriding the automatic discovery.

The Meson part already exists, although was not used by the code.

[1]: https://github.com/linuxmint/xreader/commit/9c6f363ad47f33b290d270489e50f7ad6071987a